+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_07_25_000000) do
+ActiveRecord::Schema[8.1].define(version: 2026_07_31_130000) do
   create_table "action_mailbox_inbound_emails", force: :cascade do |t|
     t.datetime "created_at", null: false
     t.string "message_checksum", null: false
@@ -166,6 +166,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_07_25_000000) do
   create_table "service_categories", force: :cascade do |t|
     t.datetime "created_at", null: false
     t.text "description"
+    t.boolean "is_hidden", default: false, null: false
     t.string "name"
     t.datetime "updated_at", null: false
   end